redirect phantom console output to note
The console.log in your phanton script is different from
the console.log in your phantom faux browser!  It is
still useful IMHO to display the phantom console.log.
Here we intercept calls to console.log, send the message
to perl.note and then call the "real" console.log
(though as far as I can tell it doesn't do anything).