Disable IOLoop timeout
If the phantomjs process is slow to respond (for whatever reason), the IOLoop's
timeout will close the connection. So any later test-result from phantomjs will
be ignored by the parent process.

This is especially a problem if no plan has been specified when calling
phantom_ok, as a single ok before the timeout will make the subtest pass.